Fix deleting lists without list-dir.

+++ lib/schleuder/list.rb
@@ -299,13 +299,18 @@ module Schleuder
     private
 
       def delete_listdir
-        if err = FileUtils.rm_r(self.listdir, secure: true)
-          logger.info "Deleted listdir"
-          return true
+        if File.exists?(self.listdir)
+          FileUtils.rm_r(self.listdir, secure: true)
+          Schleuder.logger.info "Deleted listdir"
         else
-          logger.error "Error while deleting listdir: #{err}"
-          return false
+          # Don't use list-logger here — if the list-dir isn't present we can't log to it!
+          Schleuder.logger.info "Couldn't delete listdir, directly not present"
         end
+        true
+      rescue => exc
+        # Don't use list-logger here — if the list-dir isn't present we can't log to it!
+        Schleuder.logger.error "Error while deleting listdir: #{exc}"
+        return false
       end
   end
 end
